Saturday tests

+++ inc/buffer_data.h
@@ -9,16 +9,16 @@
 
 #define SINT16_MAX (((1ull<<15)-1))
 
-extern double audio_R[RAW_BUFFER_SIZE];
-extern double audio_L[RAW_BUFFER_SIZE];
+extern float audio_R[RAW_BUFFER_SIZE];
+extern float audio_L[RAW_BUFFER_SIZE];
 
-extern double fft_re_R[CHANNEL_BUFFER_SIZE];
-extern double fft_im_R[CHANNEL_BUFFER_SIZE];
+extern float fft_re_R[CHANNEL_BUFFER_SIZE];
+extern float fft_im_R[CHANNEL_BUFFER_SIZE];
 
-extern double fft_re_L[CHANNEL_BUFFER_SIZE];
-extern double fft_im_L[CHANNEL_BUFFER_SIZE];
+extern float fft_re_L[CHANNEL_BUFFER_SIZE];
+extern float fft_im_L[CHANNEL_BUFFER_SIZE];
 
-extern double data_ILD[CHANNEL_BUFFER_SIZE];
-extern double data_IPD[CHANNEL_BUFFER_SIZE];
+extern float data_ILD[CHANNEL_BUFFER_SIZE];
+extern float data_IPD[CHANNEL_BUFFER_SIZE];
 
 #endif /* end of include guard: __HEADER_BUFFER_DATA__ */
